requirements.txt
Django==2.1.5
Pillow
djangorestframework
djangorestframework_simplejwt
pydotplus
django-extensions
graphviz
django-cors-headers
django-allauth
django-extensions
django-rest-auth
djangorestframework-jwt
boto3
boto
django-storages==1.6.6
zappa
pymysql
django-seed
pandas
numpy
drf-api-tracking
django-reset-migrations
apscheduler
django-axes